Move cli stuff to directory
