+++ src/aga/cli/__init__.py
@@ -1,0 +1,5 @@
+"""The command-line interface."""
+
+from .app import aga_app, click_object
+
+__all__ = ("aga_app", "click_object")
